+++ storybook/stories/index.tsx
@@ -10,11 +10,11 @@ import {storiesOf} from '@storybook/react-native'
 // import Welcome from './Welcome'
 import OnboardingLocation from '../../src/components/Onboarding/OnboardingLocation'
 import OnboardingAccelerometer from '../../src/components/Onboarding/OnboardingAccelerometer'
-import OnboardingSwiper from '../../src/components/Onboarding/OnboardingSwiper'
+// import OnboardingSwiper from '../../src/components/Onboarding/OnboardingSwiper'
 import OnboardingNotifications from '../../src/components/Onboarding/OnboardingNotifications'
 import OnboardingFindFriends from '../../src/components/Onboarding/OnboardingFindFriends'
-import OnboardingFindFriendsList from '../../src/components/Onboarding/OnboardingFindFriendsList'
 import {LocationWarningUI} from '../../src/components/modals/LocationWarning'
+import MockFindFriends from './FindFriendsList/MockFindFriends'
 
 function emptyFn() {
   /* noop */
@@ -22,13 +22,13 @@ function emptyFn() {
 
 storiesOf('Onboarding', module)
   // tslint:disable-next-line
-  .add('Main Swiper', () => <OnboardingSwiper log={console.log} onceStore={{flip: emptyFn}} />)
+  // .add('Main Swiper', () => <OnboardingSwiper log={console.log} onceStore={{flip: emptyFn}} />)
   .add('1 - Location', () => <OnboardingLocation onPress={emptyFn} />)
   .add('1a - Location Warning', () => <LocationWarningUI onPress={emptyFn} />)
   .add('2 - Accelerometer', () => <OnboardingAccelerometer onPress={emptyFn} />)
   .add('3 - Notifications', () => <OnboardingNotifications onPress={emptyFn} />)
   .add('4 - FindFriends', () => <OnboardingFindFriends onPress={emptyFn} onSkip={emptyFn} />)
-  .add('5 - FindFriendsList', () => <OnboardingFindFriendsList onPress={emptyFn} />)
+  .add('5 - FindFriendsList', () => <MockFindFriends />)
 
 // storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />)